requirements.txt
fastapi
uvicorn
wikipedia-api
openai
python-dotenv
requests
pytest
httpx
